+++ project.py
@@ -8,8 +8,8 @@ from datetime import datetime, timedelta
 from math import fabs
 
 from unit_converter.data_manager import load_data, add_to_log, refactor_value, save_data, zero_division_checker
-from unit_converter.data_models import UnitData
-from unit_converter.utils import validate_unit_group, get_converter_units, get_amount, resolve_aliases, parse_time_input, parse_date_input, get_seconds, format_value, calculate_leap_years, validate_date, get_days_from_month, get_index_from_month, gets_days_from_index
+from unit_converter.data_models import UnitData, ConversionData, ManageGroupData, ManageTypeData, AliasesData, ChangeBaseData
+from unit_converter.utils import print_introductory_messages, print_time_instructions, get_users_input, validate_unit_group, get_converter_units, get_amount, resolve_aliases, parse_time_input, parse_date_input, get_seconds, format_value, calculate_leap_years, validate_date, get_days_from_month, get_index_from_month, gets_days_from_index
 
 
 def main() -> None:
@@ -87,9 +87,9 @@ def handle_cli(args, data):
     elif parsed_args.command in ["history", "h"]:
         print_history(data, parsed_args.limit)
     elif parsed_args.command in ["types", "t"]:
-        validate_unit_group(parsed_args.unit_group.lower(), data)
-        unit_data = UnitData(unit_group=parsed_args.unit_group.lower())
-        print_types(data, unit_data)
+        unit_group = parsed_args.unit_group.lower()
+        validate_unit_group(unit_group, data)
+        print_types(data, unit_group)
     elif parsed_args.command in ["convert", "c"]:
         validate_unit_group(parsed_args.unit_group.lower(), data)
         unit_data = UnitData(unit_group=parsed_args.unit_group.lower())
@@ -108,7 +108,7 @@ def handle_cli(args, data):
             print(f"{format_value(unit_data.amount)} {unit_data.from_type} = {format_value(unit_data.new_value)} {unit_data.to_type}")
 
     elif parsed_args.command in ["manage-group", "mg"]:
-        validate_unit_group(parsed_args.unit_group.lower(), data, name="manage_type")
+        validate_unit_group(parsed_args.unit_group.lower(), data)
         unit_data = UnitData(
             unit_group = parsed_args.unit_group.lower(),
             action = parsed_args.action.lower(),
@@ -118,7 +118,7 @@ def handle_cli(args, data):
         manage_group(data, unit_data)
 
     elif parsed_args.command in ["manage-type", "mt"]:
-        validate_unit_group(parsed_args.unit_group.lower(), data, name="manage_type")
+        validate_unit_group(parsed_args.unit_group.lower(), data)
         unit_data = UnitData(
             unit_group = parsed_args.unit_group.lower(),
             unit_type = parsed_args.unit_type.lower(),
@@ -155,20 +155,10 @@ def handle_cli(args, data):
         change_base_unit(data, unit_data)
 
 
-def print_introductory_messages() -> None:
-    """Prints introductory messages and instructions"""
-    print("Welcome to unit converter!")
-    print("To check all group of units available, enter 'groups'")
-    print("To check all types for a specific group, enter 'types'")
-    print("To convert an unit, enter 'convert'")
-    print("To add a new unit group or a new unit type, enter 'add'")
-    print("Quit anytime by entering 'quit' or by pressing ctrl+d or ctrl+c", end="\n\n")
-
-
 def get_action(data) -> None:
     while True:
         try:
-            action: str = input("Let's begin! What do you want to do? ").strip().lower()
+            action: str = get_users_input("Let's begin! What do you want to do? ").strip().lower()
             # Adds logic to check action validity
             if action in ["groups", "g"]:
                 print_groups(data)
@@ -194,8 +184,6 @@ def get_action(data) -> None:
             elif action in ["change-base", "cb"]:
                 change_base_unit(data)
                 continue
-            elif action == "quit":
-                sys.exit("Bye!")
             else:
                 raise ValueError(f"'{action}' is not a valid action!")
         except (EOFError, KeyboardInterrupt):
@@ -207,8 +195,8 @@ def get_action(data) -> None:
 
 def get_unit_group(data) -> str:
     """Gets unit group"""
-    unit_group: str = input("Unit group: ").strip().lower()
-    validate_unit_group(data, unit_group)
+    unit_group: str = get_users_input("Unit group: ").strip().lower()
+    validate_unit_group(unit_group, data)
     return unit_group
 
 def print_groups(data) -> None:
@@ -234,11 +222,10 @@ def print_history(data, limit: int = 10) -> None:
             print(f"{format_value(entry['amount'])} {entry['from_type']} = {format_value(entry['result'])} {entry['to_type']} (Group: {entry['unit_group']})")
 
 
-def print_types(data, unit_data=None) -> None:
+def print_types(data, unit_group=None) -> None:
     """Prints all unit types for a specific unit group"""
-    if unit_data is None:
-        unit_data = UnitData(unit_group = get_unit_group(data))    
-    unit_group = unit_data.unit_group
+    if unit_group is None:
+        unit_group = get_unit_group(data)
     # Used to construct the sequence of unit_type with its respective aliases
     formatted_output = []
     # Iterates over the outter keys of 'units' dictionary
@@ -259,7 +246,7 @@ def conversion_logic(data) -> None:
     unit_data = UnitData(unit_group)
     if unit_data.unit_group == "time":
         print_time_instructions()
-        unit_data.time_input = input("Enter time conversion: ").strip().lower()
+        unit_data.time_input = get_users_input("Enter time conversion: ").strip().lower()
         unit_data.validate_time_input()
         converter_time(data, unit_data)
         return
@@ -438,28 +425,20 @@ def converter_time_2args(data, unit_data):
     unit_data.factor_time=factor_time
     add_to_log(data, unit_data, is_time_convertion=True)
 
-def print_time_instructions():
-    """Prints instructions for converting date and time units"""
-    print("For date-time conversion, you can choose from different approaches for conversion:")
-    print(" - From a specific unit to another. Usage: <unit_type> <unit_type> quantity")
-    print(" - From a more complex time to another. Usage: HH:MM:SS HH:MM:SS <unit_type>")
-    print(" - From one month to another. Usage: <month_name> <month_name> <unit_type>")
-    print(" - From a date to another. Usage: YYYY-MM-DD YYYY-MM-DD <unit_type>")
-
 
 def manage_group(data, unit_data=None) -> None:
     """Handles all logic of adding and removing unit groups"""    
     if unit_data is None:
         unit_data = UnitData(unit_group = None)
     if unit_data.action is None:
-        unit_data.action = input(f"Existed groups: {", ".join(data.units.keys())}. What do you want to do? (enter 'add' or 'remove') ").strip().lower()
+        unit_data.action = get_users_input(f"Existed groups: {", ".join(data.units.keys())}. What do you want to do? (enter 'add' or 'remove') ").strip().lower()
     unit_data.validate_action()
     if unit_data.action == "add": 
-        unit_data.unit_group = input(f"Unit group: ").strip().lower()
+        unit_data.unit_group = get_users_input(f"Unit group: ").strip().lower()
         unit_group = unit_data.unit_group
         if unit_group in data.units:
             raise ValueError(f"'{unit_group}' is already an existed group!")
-        unit_data.new_base_unit = input(f"You are creating '{unit_group}' group. Enter the base unit for that group: ").strip().lower()
+        unit_data.new_base_unit = get_users_input(f"You are creating '{unit_group}' group. Enter the base unit for that group: ").strip().lower()
         new_base_unit = unit_data.new_base_unit
         if not new_base_unit:
             raise ValueError("You need to specify the base unit to create a new group")
@@ -484,23 +463,23 @@ def manage_type(data, unit_data=None) -> None:
     if unit_data is None:
         unit_data = UnitData(unit_group = get_unit_group(data))
     if unit_data.action is None:
-        unit_data.action = input(f"Existed types for '{unit_data.unit_group}' group: {data.units[unit_data.unit_group]}. What do you want to do? (enter 'add' or 'remove') ").strip().lower()
+        unit_data.action = get_users_input(f"Existed types for '{unit_data.unit_group}' group: {data.units[unit_data.unit_group]}. What do you want to do? (enter 'add' or 'remove') ").strip().lower()
     unit_data.validate_action()
     if unit_data.action == "add":
         if unit_data.unit_type is None:
-            unit_data.unit_type = input(f"Enter new type for '{unit_data.unit_group}' group: ").strip().lower()
+            unit_data.unit_type = get_users_input(f"Enter new type for '{unit_data.unit_group}' group: ").strip().lower()
         unit_data.validate_add_action(data)
         if unit_data.unit_group == "temperature":
             add_temp_type(data, unit_data)
             return
         if unit_data.value is None:
-            unit_data.value = input(f"Enter conversion factor to base unit '{data.base_units[unit_data.unit_group]}' of '{unit_data.unit_group}' group: ").strip().lower()
+            unit_data.value = get_users_input(f"Enter conversion factor to base unit '{data.base_units[unit_data.unit_group]}' of '{unit_data.unit_group}' group: ").strip().lower()
         unit_data.validate_value()
         data.units[unit_data.unit_group][unit_data.unit_type] = unit_data.value
         print(f"A new unit type was added on '{unit_data.unit_group}' group: {unit_data.unit_type} = {unit_data.value}")
     elif unit_data.action == "remove":
         if unit_data.unit_type is None:
-            unit_data.unit_type = input(f"Enter type to remove from '{unit_data.unit_group}' group: ").strip().lower()
+            unit_data.unit_type = get_users_input(f"Enter type to remove from '{unit_data.unit_group}' group: ").strip().lower()
         unit_data.validate_remove_action(data)
         data.units[unit_data.unit_group].pop(unit_data.unit_type)
         aliases_to_remove = [alias for alias, unit in data.unit_aliases[unit_data.unit_group].items() if unit == unit_data.unit_type]
@@ -514,10 +493,10 @@ def manage_type(data, unit_data=None) -> None:
 def add_temp_type(data, unit_data) -> None:
     """Handles all logic for adding temperature units"""
     if unit_data.factor == None:
-        unit_data.factor = input(f"Enter conversion factor to base unit '{data.base_units[unit_data.unit_group]}' of 'temperature' group: ").strip().lower()
+        unit_data.factor = get_users_input(f"Enter conversion factor to base unit '{data.base_units[unit_data.unit_group]}' of 'temperature' group: ").strip().lower()
     unit_data.validate_factor()
     if unit_data.offset == None:
-        unit_data.offset = input(f"Enter offset value to base unit '{data.base_units[unit_data.unit_group]}' of 'temperature' group: ").strip().lower()
+        unit_data.offset = get_users_input(f"Enter offset value to base unit '{data.base_units[unit_data.unit_group]}' of 'temperature' group: ").strip().lower()
     unit_data.validate_offset()
     data.units[unit_data.unit_group][unit_data.unit_type] = [unit_data.factor, unit_data.offset]
     print(f"A new unit type was added on temperature group: {unit_data.unit_type} = [{unit_data.factor}, {unit_data.offset}]")
@@ -530,14 +509,14 @@ def manage_aliases(data, unit_data=None):
         unit_data = UnitData(unit_group = get_unit_group(data))
     unit_group = unit_data.unit_group
     if unit_data.unit_type is None:
-        unit_data.unit_type = resolve_aliases(data, unit_group, input(f"Enter unit type for '{unit_group}' group: ").strip().lower())
+        unit_data.unit_type = resolve_aliases(data, unit_group, get_users_input(f"Enter unit type for '{unit_group}' group: ").strip().lower())
     unit_type = unit_data.unit_type
     all_aliases = [alias for alias in data.unit_aliases[unit_group] if unit_type == data.unit_aliases[unit_group].get(alias)]
     if unit_data.action is None:
-        unit_data.action = input(f"Existed aliases for '{unit_type}': {all_aliases}. What do you want to do? (enter 'add' or 'remove') ").strip().lower()
+        unit_data.action = get_users_input(f"Existed aliases for '{unit_type}': {all_aliases}. What do you want to do? (enter 'add' or 'remove') ").strip().lower()
     action = unit_data.action
     if unit_data.alias is None:
-        unit_data.alias = input(f"Which alias do you want to {action} {"to" if action == "add" else "from"} '{unit_type}'? ").strip().lower()
+        unit_data.alias = get_users_input(f"Which alias do you want to {action} {"to" if action == "add" else "from"} '{unit_type}'? ").strip().lower()
     alias = unit_data.alias
     unit_data.validate_for_aliases(data)
     if action == "add":
@@ -557,7 +536,7 @@ def change_base_unit(data, unit_data=None):
     unit_group = unit_data.unit_group
     if unit_data.new_base_unit is None:
         print(f"All unit types for '{unit_group}' group: " + ", ".join(data.units[unit_group].keys()))
-        unit_data.new_base_unit = resolve_aliases(data, unit_group, input(f"Enter new base unit for '{unit_group}' group: ").strip().lower())
+        unit_data.new_base_unit = resolve_aliases(data, unit_group, get_users_input(f"Enter new base unit for '{unit_group}' group: ").strip().lower())
     unit_data.validate_new_base_unit(data)
 
     refactor_value(data, unit_group, unit_data.new_base_unit)